improvement: refactor files, subblock spacing
