--- sim/app/w/agents/agents.tsx
@@ -1,62 +1,0 @@
-'use client'
-
-import { useEffect } from 'react'
-import { useAgentStore } from './stores/store'
-
-/**
- * Main Agent Builder Component
- *
- * This is the root component for the Agent Builder application:
- * - Initializes and manages application state
- * - Coordinates between components
- * - Handles loading states and errors
- * - Renders the main application layout
- *
- * @module Agents
- */
-
-export default function Agents() {
-  const {
-    agents,
-    filteredAgents,
-    mcpServers,
-    loading,
-    error,
-    setLoading,
-    setError,
-    setAgents,
-    setMCPServers,
-  } = useAgentStore()
-
-  // Initialize store data on component mount
-  useEffect(() => {
-    // This would typically fetch data from an API
-    // For now, we'll just initialize with empty arrays
-    setAgents([])
-    setMCPServers([])
-    setLoading(false)
-  }, [setAgents, setMCPServers, setLoading])
-
-  if (loading) {
-    return <div>Loading agent builder...</div>
-  }
-
-  if (error) {
-    return <div>Error: {error}</div>
-  }
-
-  return (
-    <div className="agents-container">
-      <div className="agents-layout">
-        {/* Main content would be rendered here */}
-        <div className="agents-empty-state">
-          <h1>Agent Builder</h1>
-          <p>Create and manage your AI agents</p>
-          <p>Total agents: {agents.length}</p>
-          <p>Total MCP servers: {mcpServers.length}</p>
-          <p>This is a placeholder. Implement your components in the respective folders.</p>
-        </div>
-      </div>
-    </div>
-  )
-}
