--- sim/app/w/agents/components/sidebar/sidebar.tsx
@@ -1,8 +1,0 @@
-/**
- * Sidebar component for the Agent Builder
- *
- * This component provides navigation and management for agents:
- * - Lists all available agents
- * - Allows creation of new agents
- * - Handles agent selection
- */
