--- sim/app/w/agents/components/toolbar/toolbar.tsx
@@ -1,10 +1,0 @@
-/**
- * Toolbar component for the Agent Builder
- *
- * This component provides configuration tools for agents:
- * - Model selection controls
- * - MCP server connection management
- * - Prompt configuration options
- * - Testing and deployment actions
- * - Save/publish functionality
- */
