--- sim/app/w/agents/page.tsx
@@ -1,3 +1,0 @@
-import Agents from './agents'
-
-export default Agents
