--- sim/app/w/agents/utils/agent.ts
@@ -1,15 +1,0 @@
-/**
- * Agent Management Utilities
- *
- * This module provides utilities for working with agent configurations:
- * - Validation of agent configurations
- * - Creation of default agents with sensible defaults
- * - Template generation for common agent use cases
- * - ID generation and unique identifier management
- * - Configuration transformation and normalization
- *
- * These utilities help maintain consistency in agent configurations
- * and provide helper functions for common agent-related operations.
- *
- * @module AgentUtils
- */
