--- sim/app/w/agents/utils/mcp-server.ts
@@ -1,15 +1,0 @@
-/**
- * MCP Server Connection Utilities
- *
- * This module provides utilities for managing connections to MCP servers via Smithery:
- * - Connection testing and validation
- * - Connection pooling for efficient resource usage
- * - Model discovery and capability detection
- * - Agent execution on remote servers
- * - Health monitoring and status checks
- *
- * These utilities abstract the complexities of Smithery integration and provide
- * a clean API for the rest of the application to use when interacting with MCP servers.
- *
- * @module MCPServerUtils
- */
